requirements.txt
amqp==2.6.1
authy==2.2.3
Babel==2.9.1
billiard==3.6.4.0
boto==2.48.0
boto3==1.7.40
botocore==1.10.40
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.12
codeclimate-test-reporter==0.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Deprecated==1.2.13
dj-database-url==0.4.0
Django==2.2.16
django-admin-sortable2==0.7.1
django-cors-headers==2.1.0
django-extensions==3.0.9
django-phonenumber-field==2.3.1
django-redis==4.5.0
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.6.0
docopt==0.4.0
docutils==0.18.1
drf-yasg==1.20.0
entrypoints==0.3
flake8==3.7.7
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.4.5
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
kombu==4.6.11
mailchimp==2.0.9
MarkupSafe==2.1.0
mccabe==0.6.1
newrelic==4.18.0.118
openapi-codec==1.3.2
packaging==20.4
phonenumbers==8.10.11
psycopg2-binary==2.8.2
pushjack==1.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-http-client==3.3.6
pytz==2021.3
raven==6.7.0
redis==4.1.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.1.13
sendgrid==6.0.2
sentry-sdk==0.7.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stripe==2.36.1
uritemplate==4.1.1
urllib3==1.26.8
vine==1.3.0
wrapt==1.13.3
